+++ software_esp32/include/stm32.h
@@ -42,9 +42,9 @@
 #pragma once
 
 #include <Arduino.h>
-#include <SPIFFS.h>
 #include <FS.h>
 #include <CRC32.h>
+#include <LittleFS.h>
 
 
 #define     STM32OTA_START          0x12