+++ software_esp32/src/mqtt.cpp
@@ -54,6 +54,7 @@
 #include "helper.h"
 #include "web.h"
 #include "PIControl.h"
+#include "Messenger.h"
 
 
 CMqtt Mqtt;
@@ -75,7 +76,7 @@ void CMqtt::mqtt_setup(IPAddress brokerIP,uint16_t brokerPort)
         memset(lastTempValues[i].id,0x0,sizeof(lastTempValues[i].id));
         lastTempValues[i].publishNow=false;
     }
-
+    messengerSend=false;
     mqttReceived=false;
     mqtt_client.setServer(brokerIP, brokerPort);
     mqtt_client.setCallback(mcallback);
@@ -182,6 +183,17 @@ void CMqtt::mqtt_loop()
             tsPublish = millis();
             firstPublish=false;
             publish_all(check);
+        }
+        messengerSend=false;
+        connectTimeout=millis();
+    } else {
+        if ((VdmConfig.configFlash.messengerConfig.reason.reasonFlags.mqttTimeOut) && (!messengerSend)){
+            if ((millis()-connectTimeout)>((uint32_t)60*1000*VdmConfig.configFlash.messengerConfig.reason.mqttTimeOutTime)) {
+                String title = String(VdmConfig.configFlash.systemConfig.stationName) + " : MQTT" ;
+                String s = "MQTT connect failed after "+String(VdmConfig.configFlash.messengerConfig.reason.mqttTimeOutTime)+" minutes";
+                Messenger.sendMessage (title.c_str(),s.c_str());
+                messengerSend=true;
+            }  
         }
     }