+++ software_esp32/src/web.cpp
@@ -69,6 +69,28 @@ String CWeb::getSysConfig (VDM_SYSTEM_CONFIG sysConfig)
   return result;  
 }
 
+String CWeb::getMsgConfig (VDM_MSG_CONFIG msgConfig)
+{
+  String result = "{\"reason\":{\"valveBlocked\":"+String(msgConfig.reason.reasonFlags.valveBlocked)+","+
+                  "\"notDetect\":"+String(msgConfig.reason.reasonFlags.notDetect)+","+
+                  "\"reset\":"+String(msgConfig.reason.reasonFlags.reset)+","+
+                  "\"mqttTimeOut\":"+String(msgConfig.reason.reasonFlags.mqttTimeOut)+","+
+                  "\"mqttTimeOutTime\":"+String(msgConfig.reason.mqttTimeOutTime)+"},"+
+                  "\"PO\":{\"active\":"+String(msgConfig.activeFlags.pushOver)+","+
+                  "\"appToken\":\""+String(msgConfig.pushover.appToken)+"\","+
+                  "\"userToken\":\""+String(msgConfig.pushover.userToken)+"\","+
+                  "\"title\":\""+String(msgConfig.pushover.title)+"\"},"+
+                  "\"Email\":{\"active\":"+String(msgConfig.activeFlags.email)+","+
+                  "\"user\":\""+String(msgConfig.email.user)+"\","+
+                  "\"pwd\":\""+String(msgConfig.email.pwd)+"\","+
+                  "\"host\":\""+String(msgConfig.email.host)+"\","+
+                  "\"port\":\""+String(msgConfig.email.port)+"\","+
+                  "\"recipient\":\""+String(msgConfig.email.recipient)+"\","+
+                  "\"title\":\""+String(msgConfig.email.title)+"\"}"+
+                  "}";  
+  return result;  
+}
+
 String CWeb::getNetConfig (VDM_NETWORK_CONFIG netConfig)
 {
   String result = "{\"ethWifi\":"+String(netConfig.eth_wifi)+","+